src/backend/requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Uploads==0.2.1
requests==2.26.0
python-dotenv==0.19.2
firebase-admin==5.2.0
openai>=1.3.5
Werkzeug==2.1.2
uuid
Pillow==9.5.0
